% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_intervention_functions.R
\name{compose_pi}
\alias{compose_pi}
\title{Compose multiple π(t) functions multiplicatively}
\usage{
compose_pi(...)
}
\arguments{
\item{...}{Any number of functions \code{pi_fn(t)}, each in \eqn{[0,1]}.}
}
\value{
A composite function \code{pi_fn(t)} in \eqn{[0,1]}.
}
\description{
Each component function represents an independent non-pharmaceutical
intervention (NPI).  Their effects are assumed multiplicative on
transmission.
}
\examples{
lockdown  <- build_pi_step(c(10, 60), c(1, 0.4, 0.7))
masks     <- build_pi_spline(c(0, 10, 20, 30), c(1, 0.92, 0.85, 0.75))
combined  <- compose_pi(lockdown, masks)
}
