% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plotting.R
\name{EpiNova-plotting}
\alias{EpiNova-plotting}
\title{Visualisation Layer for EpiNova}
\description{
Publication-ready ggplot2 graphics. All plot functions return
ggplot2 objects that can be further customised.
}
