% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plotting.R
\name{plot_multipatch_snapshot}
\alias{plot_multipatch_snapshot}
\title{Multi-patch bar chart of infected proportion by patch}
\usage{
plot_multipatch_snapshot(multipatch_df, t_snapshot, patch_names = NULL)
}
\arguments{
\item{multipatch_df}{Data frame from \code{solve_multipatch}.}

\item{t_snapshot}{Day to visualise.}

\item{patch_names}{Character vector of patch names.}
}
\value{
A ggplot2 bar chart.
}
\description{
Multi-patch bar chart of infected proportion by patch
}
