% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plotting.R
\name{plot_scenarios}
\alias{plot_scenarios}
\title{Plot scenario comparison}
\usage{
plot_scenarios(scenario_df, obs_Y = NULL)
}
\arguments{
\item{scenario_df}{Data frame from \code{project_scenarios} or a
manually constructed data frame with columns \code{time},
\code{I_median}, \code{I_lower}, \code{I_upper}, \code{scenario}.}

\item{obs_Y}{Optional observed data to overlay.}
}
\value{
A ggplot2 object.
}
\description{
Plot scenario comparison
}
